update azure managed to one package
