--- azuremanaged/settings.gradle
@@ -1,5 +1,0 @@
-rootProject.name = 'durabletask-azuremanaged'
-
-include 'client'
-include 'worker'
-include 'shared'
