--- azuremanaged/shared/build.gradle
@@ -1,56 +1,0 @@
-plugins {
-    id 'java'
-    id 'com.google.protobuf' version '0.8.16'
-    id 'idea'
-}
-
-archivesBaseName = 'durabletask-azuremanaged-shared'
-
-def grpcVersion = '1.59.0'
-def protocVersion = '3.12.0'
-def jacksonVersion = '2.15.3'
-def azureCoreVersion = '1.45.0'
-def azureIdentityVersion = '1.11.1'
-// When build on local, you need to set this value to your local jdk11 directory.
-// Java11 is used to compile and run all the tests.
-def PATH_TO_TEST_JAVA_RUNTIME = System.env.JDK_11 ?: System.getProperty("java.home")
-
-repositories {
-    mavenCentral()
-    jcenter()
-}
-
-dependencies {
-    implementation "io.grpc:grpc-protobuf:${grpcVersion}"
-    implementation "io.grpc:grpc-stub:${grpcVersion}"
-    runtimeOnly "io.grpc:grpc-netty-shaded:${grpcVersion}"
-
-    implementation "com.azure:azure-core:${azureCoreVersion}"
-    implementation "com.azure:azure-identity:${azureIdentityVersion}"
-
-    testImplementation(platform('org.junit:junit-bom:5.7.2'))
-    testImplementation('org.junit.jupiter:junit-jupiter')
-
-    // Test dependencies
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.10.0'
-    testImplementation 'org.junit.jupiter:junit-jupiter-params:5.10.0'
-    testImplementation 'org.mockito:mockito-core:5.3.1'
-    testImplementation 'org.mockito:mockito-junit-jupiter:5.3.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
-}
-
-compileJava {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
-}
-
-compileTestJava {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
-    options.fork = true
-    options.forkOptions.executable = "${PATH_TO_TEST_JAVA_RUNTIME}/bin/javac"
-}
-
-test {
-    useJUnitPlatform()
-}
