+++ azuremanaged/src/test/java/com/microsoft/durabletask/azuremanaged/DurableTaskSchedulerClientOptionsTest.java
@@ -1,0 +1,231 @@
+package com.microsoft.durabletask.azuremanaged;
+
+import com.azure.core.credential.TokenCredential;
+import com.microsoft.durabletask.azuremanaged.DurableTaskSchedulerConnectionString;
+import io.grpc.Channel;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+import java.time.Duration;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+/**
+ * Unit tests for {@link DurableTaskSchedulerClientOptions}.
+ */
+@ExtendWith(MockitoExtension.class)
+public class DurableTaskSchedulerClientOptionsTest {
+
+    private static final String VALID_ENDPOINT = "https://example.com";
+    private static final String VALID_TASKHUB = "myTaskHub";
+    private static final String VALID_CONNECTION_STRING = 
+        "Endpoint=https://example.com;Authentication=ManagedIdentity;TaskHub=myTaskHub";
+
+    @Mock
+    private TokenCredential mockCredential;
+
+    @Test
+    @DisplayName("Default constructor should set default values")
+    public void defaultConstructor_SetsDefaultValues() {
+        // Act
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions();
+        
+        // Assert
+        assertEquals("", options.getEndpointAddress());
+        assertEquals("", options.getTaskHubName());
+        assertNull(options.getCredential());
+        assertEquals("https://durabletask.io", options.getResourceId());
+        assertFalse(options.isAllowInsecureCredentials());
+        assertEquals(Duration.ofMinutes(5), options.getTokenRefreshMargin());
+    }
+
+    @Test
+    @DisplayName("fromConnectionString should parse connection string correctly")
+    public void fromConnectionString_ParsesConnectionStringCorrectly() {
+        // Act
+        DurableTaskSchedulerClientOptions options = 
+            DurableTaskSchedulerClientOptions.fromConnectionString(VALID_CONNECTION_STRING, mockCredential);
+        
+        // Assert
+        assertEquals(VALID_ENDPOINT, options.getEndpointAddress());
+        assertEquals(VALID_TASKHUB, options.getTaskHubName());
+        assertSame(mockCredential, options.getCredential());
+        assertFalse(options.isAllowInsecureCredentials());
+    }
+
+    @Test
+    @DisplayName("fromConnectionString should allow insecure credentials when credential is null")
+    public void fromConnectionString_AllowsInsecureCredentialsWhenCredentialIsNull() {
+        // Act
+        DurableTaskSchedulerClientOptions options = 
+            DurableTaskSchedulerClientOptions.fromConnectionString(VALID_CONNECTION_STRING, null);
+        
+        // Assert
+        assertTrue(options.isAllowInsecureCredentials());
+        assertNull(options.getCredential());
+    }
+
+    @Test
+    @DisplayName("setEndpointAddress should update endpoint address")
+    public void setEndpointAddress_UpdatesEndpointAddress() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions();
+        
+        // Act
+        DurableTaskSchedulerClientOptions result = options.setEndpointAddress(VALID_ENDPOINT);
+        
+        // Assert
+        assertEquals(VALID_ENDPOINT, options.getEndpointAddress());
+        assertSame(options, result); // Builder pattern returns this
+    }
+
+    @Test
+    @DisplayName("setTaskHubName should update task hub name")
+    public void setTaskHubName_UpdatesTaskHubName() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions();
+        
+        // Act
+        DurableTaskSchedulerClientOptions result = options.setTaskHubName(VALID_TASKHUB);
+        
+        // Assert
+        assertEquals(VALID_TASKHUB, options.getTaskHubName());
+        assertSame(options, result); // Builder pattern returns this
+    }
+
+    @Test
+    @DisplayName("setCredential should update credential")
+    public void setCredential_UpdatesCredential() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions();
+        
+        // Act
+        DurableTaskSchedulerClientOptions result = options.setCredential(mockCredential);
+        
+        // Assert
+        assertSame(mockCredential, options.getCredential());
+        assertSame(options, result); // Builder pattern returns this
+    }
+
+    @Test
+    @DisplayName("setResourceId should update resource ID")
+    public void setResourceId_UpdatesResourceId() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions();
+        String customResourceId = "https://custom-resource.example.com";
+        
+        // Act
+        DurableTaskSchedulerClientOptions result = options.setResourceId(customResourceId);
+        
+        // Assert
+        assertEquals(customResourceId, options.getResourceId());
+        assertSame(options, result); // Builder pattern returns this
+    }
+
+    @Test
+    @DisplayName("setAllowInsecureCredentials should update allowInsecureCredentials")
+    public void setAllowInsecureCredentials_UpdatesAllowInsecureCredentials() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions();
+        
+        // Act
+        DurableTaskSchedulerClientOptions result = options.setAllowInsecureCredentials(true);
+        
+        // Assert
+        assertTrue(options.isAllowInsecureCredentials());
+        assertSame(options, result); // Builder pattern returns this
+    }
+
+    @Test
+    @DisplayName("setTokenRefreshMargin should update token refresh margin")
+    public void setTokenRefreshMargin_UpdatesTokenRefreshMargin() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions();
+        Duration customMargin = Duration.ofMinutes(10);
+        
+        // Act
+        DurableTaskSchedulerClientOptions result = options.setTokenRefreshMargin(customMargin);
+        
+        // Assert
+        assertEquals(customMargin, options.getTokenRefreshMargin());
+        assertSame(options, result); // Builder pattern returns this
+    }
+
+    @Test
+    @DisplayName("validate should not throw when options are valid")
+    public void validate_DoesNotThrowWhenOptionsAreValid() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions()
+            .setEndpointAddress(VALID_ENDPOINT)
+            .setTaskHubName(VALID_TASKHUB);
+        
+        // Act & Assert
+        assertDoesNotThrow(() -> options.validate());
+    }
+
+    @Test
+    @DisplayName("validate should throw when endpoint address is null")
+    public void validate_ThrowsWhenEndpointAddressIsNull() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions()
+            .setEndpointAddress(null)
+            .setTaskHubName(VALID_TASKHUB);
+        
+        // Act & Assert
+        assertThrows(NullPointerException.class, () -> options.validate());
+    }
+
+    @Test
+    @DisplayName("validate should throw when task hub name is null")
+    public void validate_ThrowsWhenTaskHubNameIsNull() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions()
+            .setEndpointAddress(VALID_ENDPOINT)
+            .setTaskHubName(null);
+        
+        // Act & Assert
+        assertThrows(NullPointerException.class, () -> options.validate());
+    }
+
+    @Test
+    @DisplayName("createGrpcChannel should create a channel")
+    public void createGrpcChannel_CreatesChannel() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions()
+            .setEndpointAddress(VALID_ENDPOINT)
+            .setTaskHubName(VALID_TASKHUB);
+        
+        // Act
+        Channel channel = options.createGrpcChannel();
+        
+        // Assert
+        assertNotNull(channel);
+    }
+
+    @Test
+    @DisplayName("createGrpcChannel should handle endpoints without protocol")
+    public void createGrpcChannel_HandlesEndpointsWithoutProtocol() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions()
+            .setEndpointAddress("example.com")
+            .setTaskHubName(VALID_TASKHUB);
+        
+        // Act & Assert
+        assertDoesNotThrow(() -> options.createGrpcChannel());
+    }
+
+    @Test
+    @DisplayName("createGrpcChannel should throw for invalid URLs")
+    public void createGrpcChannel_ThrowsForInvalidUrls() {
+        // Arrange
+        DurableTaskSchedulerClientOptions options = new DurableTaskSchedulerClientOptions()
+            .setEndpointAddress("invalid:url")
+            .setTaskHubName(VALID_TASKHUB);
+        
+        // Act & Assert
+        assertThrows(IllegalArgumentException.class, () -> options.createGrpcChannel());
+    }
+}
