+++ samples/build.gradle
@@ -7,7 +7,6 @@ plugins {
 group 'io.durabletask'
 version = '0.1.0'
 def grpcVersion = '1.59.0'
-def durabletaskClientVersion = '1.5.0'
 archivesBaseName = 'durabletask-samples'
 
 application {
@@ -16,10 +15,8 @@ application {
 
 dependencies {
     implementation project(':client')
-    implementation project(':azuremanaged:dtclient')
-    implementation project(':azuremanaged:dtworker')
+    implementation project(':azuremanaged')
 
-    // implementation "com.microsoft:durabletask-client:${durabletaskClientVersion}"
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation platform("org.springframework.boot:spring-boot-dependencies:2.5.2")
     implementation 'org.springframework.boot:spring-boot-starter'
