+++ settings.gradle
@@ -3,9 +3,6 @@ rootProject.name = 'durabletask-java'
 include ":client"
 include ":azurefunctions"
 include ":azuremanaged"
-include ":azuremanaged:dtclient"
-include ":azuremanaged:dtworker"
-include ":azuremanaged:shared"
 include ":samples"
 include ":samples-azure-functions"
 include ":endtoendtests"
